+++ src/UnityFx.Async/Api/Core/AsyncResult.Events.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.Threading;
 
@@ -47,7 +48,7 @@ namespace UnityFx.Async
 		#region IAsyncOperationEvents
 
 		/// <inheritdoc/>
-		public event AsyncOperationCallback Completed
+		public event AsyncCompletedEventHandler Completed
 		{
 			add
 			{
@@ -491,7 +492,7 @@ namespace UnityFx.Async
 				}
 				else
 				{
-					AsyncContinuation.InvokeInline(this, continuation, false);
+					InvokeContinuationInline(continuation, false);
 				}
 			}
 		}
@@ -504,11 +505,11 @@ namespace UnityFx.Async
 			}
 			else if (syncContext == null || syncContext == SynchronizationContext.Current)
 			{
-				AsyncContinuation.InvokeInline(this, continuation, true);
+				InvokeContinuationInline(continuation, true);
 			}
 			else
 			{
-				syncContext.Post(args => AsyncContinuation.InvokeInline(this, args, true), continuation);
+				syncContext.Post(args => InvokeContinuationInline(args, true), continuation);
 			}
 		}
 
@@ -516,12 +517,17 @@ namespace UnityFx.Async
 		{
 			if (syncContext != null && syncContext.GetType() != typeof(SynchronizationContext))
 			{
-				syncContext.Post(args => AsyncContinuation.InvokeInline(this, args, inline), continuation);
+				syncContext.Post(args => InvokeContinuationInline(args, inline), continuation);
 			}
 			else
 			{
-				ThreadPool.QueueUserWorkItem(args => AsyncContinuation.InvokeInline(this, args, inline), continuation);
+				ThreadPool.QueueUserWorkItem(args => InvokeContinuationInline(args, inline), continuation);
 			}
+		}
+
+		private void InvokeContinuationInline(object continuation, bool inline)
+		{
+			AsyncContinuation.InvokeInline(this, continuation, inline);
 		}
 
 #endif