+++ src/UnityFx.Async/Api/Interfaces/IAsyncOperationEvents.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using System;
+using System.ComponentModel;
 using System.Threading;
 
 namespace UnityFx.Async
@@ -16,7 +17,7 @@ namespace UnityFx.Async
 	public delegate void AsyncOperationCallback(IAsyncOperation op);
 
 	/// <summary>
-	/// A controller for <see cref="IAsyncOperation"/> completion callbacks.
+	/// Manages events and callbacks of <see cref="IAsyncOperation"/>.
 	/// </summary>
 	/// <seealso cref="IAsyncOperation"/>
 	public interface IAsyncOperationEvents
@@ -34,7 +35,7 @@ namespace UnityFx.Async
 		/// <seealso cref="TryAddContinuation(AsyncOperationCallback)"/>
 		/// <seealso cref="TryAddContinuation(AsyncOperationCallback, SynchronizationContext)"/>
 		/// <seealso cref="RemoveContinuation(AsyncOperationCallback)"/>
-		event AsyncOperationCallback Completed;
+		event AsyncCompletedEventHandler Completed;
 
 		/// <summary>
 		/// Adds a completion callback to be executed after the operation has completed. If the operation is already completed the <paramref name="action"/> is called synchronously.
@@ -114,7 +115,7 @@ namespace UnityFx.Async
 		bool RemoveContinuation(AsyncOperationCallback action);
 
 		/// <summary>
-		/// Adds a completion callback to be executed after the operation has completed. If the operation is completed <paramref name="continuation"/> is invoked synchronously.
+		/// Adds a continuation to be executed after the operation has completed. If the operation is completed <paramref name="continuation"/> is invoked synchronously.
 		/// </summary>
 		/// <remarks>
 		/// The <paramref name="continuation"/> is invoked on a thread that registered the continuation (if it has a <see cref="SynchronizationContext"/> attached).
@@ -129,7 +130,7 @@ namespace UnityFx.Async
 		void AddContinuation(IAsyncContinuation continuation);
 
 		/// <summary>
-		/// Attempts to add a completion callback to be executed after the operation has finished. If the operation is already completed
+		/// Attempts to add a continuation to be executed after the operation has finished. If the operation is already completed
 		/// the method does nothing and just returns <see langword="false"/>.
 		/// </summary>
 		/// <remarks>
@@ -146,7 +147,7 @@ namespace UnityFx.Async
 		bool TryAddContinuation(IAsyncContinuation continuation);
 
 		/// <summary>
-		/// Adds a completion callback to be executed after the operation has completed. If the operation is completed <paramref name="continuation"/> is invoked synchronously.
+		/// Adds a continuation to be executed after the operation has completed. If the operation is completed <paramref name="continuation"/> is invoked synchronously.
 		/// </summary>
 		/// <remarks>
 		/// The <paramref name="continuation"/> is invoked on a <see cref="SynchronizationContext"/> specified. Throwing an exception from the callback might cause unspecified behaviour.
@@ -163,7 +164,7 @@ namespace UnityFx.Async
 		void AddContinuation(IAsyncContinuation continuation, SynchronizationContext syncContext);
 
 		/// <summary>
-		/// Attempts to add a completion callback to be executed after the operation has finished. If the operation is already completed
+		/// Attempts to add a continuation to be executed after the operation has finished. If the operation is already completed
 		/// the method does nothing and just returns <see langword="false"/>.
 		/// </summary>
 		/// <remarks>
@@ -182,7 +183,7 @@ namespace UnityFx.Async
 		bool TryAddContinuation(IAsyncContinuation continuation, SynchronizationContext syncContext);
 
 		/// <summary>
-		/// Removes an existing completion callback.
+		/// Removes an existing continuation.
 		/// </summary>
 		/// <param name="continuation">The continuation to remove. Can be <see langword="null"/>.</param>
 		/// <returns>Returns <see langword="true"/> if <paramref name="continuation"/> was removed; <see langword="false"/> otherwise.</returns>